# admin-dashboard/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Admin Dashboard API
  description: 塗装業フランチャイズネットワーク本部管理システムAPI
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.admin-dashboard.com/v1
    description: Production server
paths:
...
  /cases:
    get:
      summary: 案件一覧取得
      tags: [Cases]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETED, CANCELLED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: Cases list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasesList'

    post:
      summary: 新規案件作成
      tags: [Cases]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        201:
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
  /cases/{id}:
    get:
      summary: 案件詳細取得
      tags: [Cases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'

    put:
      summary: 案件更新
      tags: [Cases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        200:
          description: Case updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
...
  /cases/{id}/status:
    put:
      summary: 本部ステータス更新
      tags: [Cases, Status]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hqStatus]
              properties:
                hqStatus:
                  type: string
                  enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETED, CANCELLED]
                note:
                  type: string
      responses:
        200:
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
...
components:
  schemas:
    Case:
      type: object
      properties:
        id:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        address:
          type: string
        buildingType:
          type: string
          enum: [detached, apartment]
        description:
          type: string
        estimatedValue:
          type: number
        urgency:
          type: string
          enum: [low, medium, high]
        hqStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCaseRequest:
      type: object
      required: [customerName, customerPhone, address, buildingType, description]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        address:
          type: string
        buildingType:
          type: string
          enum: [detached, apartment]
        description:
          type: string
        estimatedValue:
          type: number
        urgency:
          type: string
          enum: [low, medium, high]
    UpdateCaseRequest:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        address:
          type: string
        buildingType:
          type: string
        description:
          type: string
        estimatedValue:
          type: number
        urgency:
          type: string
    CasesList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Case'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []